Split Polisher::Gemfile specs along module boundaries
After this all specs are now organized in the same heirarchy
of what they are testing
